@@ -28,10 +28,10 @@ int main(int argc, char **argv)
     MainScene mainScene(canvasRect, windowed);
 
     std::string markup = "The quick brown fox";
-    auto label = std::unique_ptr<LabelActor>(new LabelActor(Rect(0, 0, 600, 200), markup));
+    auto label = std::make_unique<LabelActor>(Rect {0, 0, 600, 200}, markup);
     mainScene.add_actor(std::move(label));
 
-    auto qube = std::unique_ptr<QubeActor>(new QubeActor(Rect(0, 25, 300, 300)));
+    auto qube = std::make_unique<QubeActor>(Rect(0, 25, 300, 300));
     mainScene.add_actor(std::move(qube));
 
     bool running = true;
